requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
google-generativeai
Jinja2